Added styles for the initial fruit list

+++ app/src/main/java/com/example/fruitapp/FruitInfo.kt
@@ -1,3 +1,3 @@
 package com.example.fruitapp
 
-data class FruitInfo(val type: String, val price: Int, val weight: Int)
+data class FruitInfo(val type: String, val price: Float, val weight: Float)
